feat: sebulba folder restructure
