+++ mava/systems/sac/anakin/ff_isac.py
@@ -602,7 +602,7 @@ def run_experiment(cfg: DictConfig) -> float:
     return eval_performance
 
 
-@hydra.main(config_path="../../configs", config_name="default_ff_isac.yaml", version_base="1.2")
+@hydra.main(config_path="../../../configs", config_name="default_ff_isac.yaml", version_base="1.2")
 def hydra_entry_point(cfg: DictConfig) -> float:
     """Experiment entry point."""
     # Allow dynamic attributes.
