+++ test/integration_test.py
@@ -23,9 +23,14 @@ from omegaconf import DictConfig, OmegaConf
 # system run all envs, but each env and each system is run at least once.
 # For each system we select a random environment to run.
 # Then for each environment we select a random system to run.
-ppo_systems = ["ppo.ff_ippo", "ppo.ff_mappo", "ppo.rec_ippo", "ppo.rec_mappo"]
-q_learning_systems = ["q_learning.rec_iql"]
-sac_systems = ["sac.ff_isac", "sac.ff_masac"]
+ppo_systems = [
+    "ppo.anakin.ff_ippo",
+    "ppo.anakin.ff_mappo",
+    "ppo.anakin.rec_ippo",
+    "ppo.anakin.rec_mappo",
+]
+q_learning_systems = ["q_learning.anakin.rec_iql"]
+sac_systems = ["sac.anakin.ff_isac", "sac.anakin.ff_masac"]
 
 discrete_envs = ["gigastep", "lbf", "matrax", "rware", "smax"]
 cnn_envs = ["cleaner", "connector"]
@@ -59,7 +64,7 @@ def _get_fast_config(cfg: DictConfig, fast_config: dict) -> DictConfig:
 @pytest.mark.parametrize("system_path", ppo_systems)
 def test_ppo_system(fast_config: dict, system_path: str) -> None:
     """Test all ppo systems on random envs."""
-    _, system_name = system_path.split(".")
+    _, _, system_name = system_path.split(".")
     env = random.choice(discrete_envs)
 
     with initialize(version_base=None, config_path="../mava/configs/"):
@@ -72,7 +77,7 @@ def test_ppo_system(fast_config: dict, system_path: str) -> None:
 @pytest.mark.parametrize("system_path", q_learning_systems)
 def test_q_learning_system(fast_config: dict, system_path: str) -> None:
     """Test all Q-Learning systems on random envs."""
-    _, system_name = system_path.split(".")
+    _, _, system_name = system_path.split(".")
     env = random.choice(discrete_envs)
 
     with initialize(version_base=None, config_path="../mava/configs/"):
@@ -85,7 +90,7 @@ def test_q_learning_system(fast_config: dict, system_path: str) -> None:
 @pytest.mark.parametrize("system_path", sac_systems)
 def test_sac_system(fast_config: dict, system_path: str) -> None:
     """Test all SAC systems on random envs."""
-    _, system_name = system_path.split(".")
+    _, _, system_name = system_path.split(".")
     env = random.choice(continuous_envs)
 
     with initialize(version_base=None, config_path="../mava/configs/"):
@@ -99,7 +104,7 @@ def test_sac_system(fast_config: dict, system_path: str) -> None:
 def test_discrete_env(fast_config: dict, env_name: str) -> None:
     """Test all discrete envs on random systems."""
     system_path = random.choice(ppo_systems + q_learning_systems)
-    _, system_name = system_path.split(".")
+    _, _, system_name = system_path.split(".")
 
     with initialize(version_base=None, config_path="../mava/configs/"):
         cfg = compose(config_name=f"default_{system_name}", overrides=[f"env={env_name}"])
@@ -112,7 +117,7 @@ def test_discrete_env(fast_config: dict, env_name: str) -> None:
 def test_discrete_cnn_env(fast_config: dict, env_name: str) -> None:
     """Test all 2D envs on random systems."""
     system_path = random.choice(ppo_systems)
-    _, system_name = system_path.split(".")
+    _, _, system_name = system_path.split(".")
 
     network = "cnn" if "ff" in system_name else "rcnn"
 
@@ -130,7 +135,7 @@ def test_discrete_cnn_env(fast_config: dict, env_name: str) -> None:
 def test_continuous_env(fast_config: dict, env_name: str) -> None:
     """Test all continuous envs on random systems."""
     system_path = random.choice(ppo_systems + sac_systems)
-    _, system_name = system_path.split(".")
+    _, _, system_name = system_path.split(".")
 
     overrides = [f"env={env_name}", "network=continuous_mlp"]
     with initialize(version_base=None, config_path="../mava/configs/"):
